+++ lib/compony/component.rb
@@ -137,6 +137,15 @@ module Compony
       @content_blocks.natural_push(name, block || :missing, before:, **kwargs)
     end
 
+    # DSL method
+    # Removes a content block. Use this in subclasses if a content block defined in the parent should be removed from the child.
+    # @param [Symbol,String] name Name of the content block that should be removed
+    def remove_content(name)
+      name = name.to_sym
+      existing_index = @content_blocks.find_index { |el| el.name == name } || fail("Content block #{name.inspect} not found for removal in #{inspect}.")
+      @content_blocks.delete_at(existing_index)
+    end
+
     # Renders the component using the controller passsed to it and returns it as a string.
     # @param [Boolean] standalone pass true iff `render` is called from `render_standalone`
     # Do not overwrite.